requirements.txt
altgraph==0.17
click==7.1.2
cycler==0.10.0
fbs==0.9.6
future==0.18.2
kiwisolver==1.3.1
lxml==4.6.3
macholib==1.14
matplotlib==3.3.4
numpy==1.19.5
pefile==2021.5.24
Pillow==8.3.0
PyInstaller==3.4
pyparsing==2.4.7
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.2
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.18.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
six==1.16.0
